Refactored testing
The regression testing has been entirely re-written to use the python3
unittest module and to depend on as little reference data as possible
(previously tklbam compared presummed correct output with current
output to gauge correctness).

A description of the new tests and how they correlate with the old
regression tests can be found in ``utests/readme.rst``

It should be noted however the following tests have NOT been recreated

- "4 - fixstat simulation" fixstat is being tested but not the simulation
- "5 - fixstat simulation with limitation" fixstat is being tested but not the simulation
- "6 - fixstat simulation with exclusion" fixstat is being tested but not the simulation
- "11 - delete simulation", delete is tested but not it's simulation
- "12 - delete simulation with limitation", delete is tested but not it's simulation

Mysql related tests should also be fleshed out, and postgresql tests which
never existed should be added.